refactor(models): rename base-model to model
